refactor(ui): split api into crud/eval and consolidate evaluation helpers
Rename flags.ts → crud.ts and evaluation.ts → eval.ts to mirror swagger
grouping. Merge evalParse, evalDebugLog, and jsonEditorValue into
helpers/evaluation.ts. Fold flags list snapshot cache into flagsListPage.ts.

Verified: make flagr-ui-check, make test-e2e (30/30).
